Hi
The falling blocks work - again. The blocks also have their different
lanes and they don’t have a random X - int. They have a width of 35.
Make them stack on top of each other, and use arrays to do that. Also
make the player die when the blocks hit it, and make a the game over
screen, the instructions page, and the Welcome screen. Make the
transitions for each one, and don’t let them overlap. finish the game
soon, now i am typing random things because it is fun, have a good day,
this is so fun typing random things :) I got to go before I like my
fingers and die :(